% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aggregate_daily_zoo_to_monthly_ts}
\alias{aggregate_daily_zoo_to_monthly_ts}
\title{Convert a daily zoo object to a monthly \code{ts} object}
\usage{
aggregate_daily_zoo_to_monthly_ts(
  zoo_obj,
  var = "Temp",
  na.rm = TRUE,
  na_prop_max = 1
)
}
\arguments{
\item{zoo_obj}{A \code{zoo} object with daily observations.
The index must be of class \code{Date} or \code{POSIXt}.}

\item{var}{A character string specifying the name of the variable
to be aggregated (default: \code{"Temp"}).}

\item{na.rm}{Logical; should missing values be removed before averaging?}

\item{na_prop_max}{Maximum allowed proportion of NA values within a month.
If the proportion of missing values exceeds this threshold, the monthly
mean is set to NA. Default is \code{1} (no additional filtering).}
}
\value{
A monthly \code{ts} object with frequency = 12.
}
\description{
Convert a daily zoo object to a monthly \code{ts} object
}
\examples{
\dontrun{
sst_zoo <- get_jma_sst_zoo(sea_area_id = 138)
sst_ts_monthly <- aggregate_daily_zoo_to_monthly_ts(sst_138_zoo)
}

}
