% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{autoplot_season}
\alias{autoplot_season}
\title{Plot the estimated seasonal component from a tempssm model}
\usage{
autoplot_season(
  res,
  ci = TRUE,
  ci_level = 0.95,
  ylab = "Temperature",
  show_ci_in_title = FALSE
)
}
\arguments{
\item{res}{An object of class \code{"tempssm"} returned by \code{tempssm()}.}

\item{ci}{Logical; if TRUE, pointwise confidence intervals are returned.}

\item{ci_level}{Numeric confidence level between 0 and 1 (default: 0.95).}

\item{ylab}{Character string giving label of y-axis.
Defalut is "temperature".}

\item{show_ci_in_title}{Logical; should the confidence level be shown in
the plot title when \code{ci = TRUE}? The default is \code{FALSE}.}
}
\value{
A \code{ggplot} object, allowing further customization by adding
standard \pkg{ggplot2} layers.
}
\description{
Create a \pkg{ggplot2} visualization of the estimated drift component
obtained from a state space model fitted by \code{tempssm()}.
A pointwise confidence interval is shown as a shaded ribbon.
}
\details{
The confidence interval is computed using
\code{stats::confint()} applied to the Kalman filter and smoother
results stored in \code{res$kfs}. The shaded ribbon represents
pointwise confidence intervals for the level state.
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)

# Default 95\% confidence interval
autoplot_season(res)

# Custom confidence level
autoplot_season(res, ci_level = 0.9)
}
}
\seealso{
\code{\link{tempssm}}, \code{\link[stats]{confint}}
}
