% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{check_exo_ts_lgssm}
\alias{check_exo_ts_lgssm}
\title{Check ts object of exogenous variable(s) for applying \code{lgssm()}}
\usage{
check_exo_ts_lgssm(temp_data, exo_data)
}
\arguments{
\item{temp_data}{A temperature time series of class \code{ts}.
The series can have any arbitrary frequency of 2 or higher.
For example, a frequency of 12 represents a monthly \code{ts} object.}

\item{exo_data}{A data set of exogenous variable(s) of class \code{ts}.
The series may have any arbitrary frequency of 2 or higher,
but it must be the same as that of \code{temp_data}.
The default is \code{NULL} when fitting a model without exogenous variables.}
}
\value{
A univariate or multivairate \code{ts} object.
}
\description{
Check ts object of exogenous variable(s) for applying \code{lgssm()}
}
