% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2ts.R
\encoding{UTF-8}
\name{csv2ts}
\alias{csv2ts}
\title{Function to convert from data.csv to time sieries object}
\usage{
csv2ts(csv)
}
\arguments{
\item{csv}{input csv file}
}
\description{
Function to convert from data.csv to time sieries object
}
