% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_cross_validation.R
\name{.compute_mae}
\alias{.compute_mae}
\title{Compute mean absolute error (MAE)}
\usage{
.compute_mae(y_pred, y_true)
}
\arguments{
\item{y_pred}{Predicted values.}

\item{y_true}{Observed values.}
}
\value{
Numeric scalar giving MAE.
}
\description{
Compute mean absolute error (MAE)
}
