% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{.ssm_check_temp_ts}
\alias{.ssm_check_temp_ts}
\title{Check ts object of temperature time series for applying \code{ssm()}}
\usage{
.ssm_check_temp_ts(temp_data)
}
\arguments{
\item{temp_data}{A temperature time series of class \code{ts}.
The series can have any arbitrary frequency of 2 or higher.
For example, a frequency of 12 represents a monthly \code{ts} object.}
}
\value{
A univariate \code{ts} object.
}
\description{
Check ts object of temperature time series for applying \code{ssm()}
}
