% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_cross_validation.R
\name{exec_tsCV}
\alias{exec_tsCV}
\title{Execute cross-validation of each of rolling-origin train/test data}
\usage{
exec_tsCV(folds, idx = 1, ar_order_given)
}
\arguments{
\item{folds}{A \code{list} of folds object returned by \code{ts_cv_folds()}.}

\item{idx}{A \code{vector} of sequential IDs of folds for ts cross-validation.}

\item{ar_order_given}{Integer specifying the order of the autoregressive (AR)
component in the error structure (e.g., 2 for AR(2), 3 for AR(3))
for applying \code{ssm()}.}
}
\value{
A \code{list} object of ts cross-validation output
}
\description{
This function predict with using test data for in ts cross-validation.
}
