% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\encoding{UTF-8}
\name{extract_drift}
\alias{extract_drift}
\title{Function to extract smoothing estimate of drift component as ts object}
\usage{
extract_drift(res)
}
\arguments{
\item{res}{return of lgssm}
}
\description{
Function to extract smoothing estimate of drift component as ts object
}
