% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{extract_drift_ts}
\alias{extract_drift_ts}
\title{Extract smoothed drift (slope) component as a time series}
\usage{
extract_drift_ts(res)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}
}
\value{
A \code{ts} object of the smoothed drift component.
}
\description{
Extract smoothed drift (slope) component as a time series
}
