% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{extract_exo_coef_ci}
\alias{extract_exo_coef_ci}
\title{Extract coefficient of exogenous variables with confidence interval}
\usage{
extract_exo_coef_ci(res, level = 0.95)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}

\item{level}{Percentage of confidence interval. The default value of 0.95.}
}
\value{
A \code{data.frame} object of CI for exogenous variable(s).
}
\description{
Extract coefficient of exogenous variables with confidence interval
}
