% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\encoding{UTF-8}
\name{extract_level}
\alias{extract_level}
\title{Function to extract smoothing estimate of level component as ts object}
\usage{
extract_level(res)
}
\arguments{
\item{res}{return of lgssm}
}
\description{
Function to extract smoothing estimate of level component as ts object
}
