% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{extract_level_ts}
\alias{extract_level_ts}
\title{Extract smoothed level component as a time series}
\usage{
extract_level_ts(res)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}
}
\value{
A \code{ts} object of the smoothed level component.
}
\description{
Extract smoothed level component as a time series
}
