% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{extract_param}
\alias{extract_param}
\title{Extract estimated parameters in the fitted models}
\usage{
extract_param(res)
}
\arguments{
\item{res}{An object of class \code{"tempssm"} returned by \code{lgssm()}.}
}
\value{
A \code{list} object of the estimated parameters.
}
\description{
Extract estimated parameters in the fitted models
}
