% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\encoding{UTF-8}
\name{extract_pars}
\alias{extract_pars}
\title{Function to extract process/observation error parameters}
\usage{
extract_pars(res)
}
\arguments{
\item{res}{return of lgssm}
}
\description{
Function to extract process/observation error parameters
}
