% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{extract_regression_ci}
\alias{extract_regression_ci}
\title{Extract CI of coefficient for exogenous variable(s)}
\usage{
extract_regression_ci(res, level = 0.95)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}
}
\value{
A \code{data.frame} object of CI for exogenous variable(s).
}
\description{
Extract CI of coefficient for exogenous variable(s)
}
