% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_temp_dev}
\alias{generate_temp_dev}
\title{Generate a monthly temperature anomaly time series}
\usage{
generate_temp_dev(temp_ts)
}
\arguments{
\item{temp_ts}{A monthly temperature time series object of class \code{ts}.}
}
\value{
A \code{ts} object of monthly temperature anomalies.
}
\description{
Temperature anomalies are calculated by subtracting the long-term
monthly mean (climatology) from each observation.
}
\details{
Monthly climatology is calculated as the mean temperature for each calendar
month. Missing values are ignored.
}
\examples{
temp_ts <- ts(
  c(5.2, 6.1, 9.3, 13.5, 17.8, 21.0,
    24.3, 25.1, 22.0, 17.1, 11.2, 7.0),
  start = c(2000, 1),
  frequency = 12
)

temp_dev <- generate_temp_dev(temp_ts)
tapply(as.numeric(temp_dev), cycle(temp_dev), mean)

}
