% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagonostics.R
\name{get_residuals}
\alias{get_residuals}
\title{Extract standardized recursive residuals}
\usage{
get_residuals(res)
}
\arguments{
\item{res}{An object of class "tempssm".}
}
\value{
A numeric vector of standardized recursive residuals.
}
\description{
Extract standardized recursive residuals
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)

residuals <- get_residuals(res)
}
}
