% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempssm_accessors.R
\name{get_tempssm_params}
\alias{get_tempssm_params}
\title{Extract estimated parameters in the fitted models}
\usage{
get_tempssm_params(res)
}
\arguments{
\item{res}{An object of class \code{"tempssm"} returned by \code{tempssm()}.}
}
\value{
A \code{list} object of the estimated parameters.
}
\description{
Extract estimated parameters in the fitted models
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)
params <- get_tempssm_params(res)
}
}
