% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibaraki_sst.R
\docType{data}
\name{ibaraki_sst}
\alias{ibaraki_sst}
\title{Daily mean temperature time series off southern Ibaraki Prefecture}
\format{
A \code{zoo} object with:
\describe{
  \item{index}{Date from 1982-01-01 to 2025-12-31}
  \item{Temp}{Daily mean temperature (°C)}
}
}
\source{
Japan Meteorological Agency (JMA). Data obtained from the JMA website:
\url{https://www.jma.go.jp/jma/indexe.html}
}
\usage{
ibaraki_sst
}
\description{
A daily mean temperature time series observed in the coastal waters
off southern Ibaraki Prefecture, Japan. The data are provided as a
\code{zoo} object and serve as an example dataset for state-space
analysis of daily temperature time series.
}
\details{
The variable name is unified as \code{Temp} for consistency with other
example datasets in the package. In this dataset, \code{Temp}
represents daily mean sea surface temperature. The observation period
spans from January 1, 1982 to December 31, 2025. Missing values, if any,
are encoded as \code{NA}.
}
\examples{
data(ibaraki_sst)
plot(ibaraki_sst,
     ylab = "Temperature (°C)",
     main = "Daily mean sea surface temperature off southern Ibaraki Prefecture")

}
\keyword{datasets}
