% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{label_ts_mono}
\alias{label_ts_mono}
\title{Assign a column label to a univariate \code{ts} object}
\usage{
label_ts_mono(ts_in, label = "var")
}
\arguments{
\item{ts_in}{A univariate time series object of class \code{ts}.}

\item{label}{A character string specifying the column name to assign
(default: \code{"var"}).}
}
\value{
A \code{ts} object with a single column labeled by \code{label}.
}
\description{
This function assigns a column name to a univariate time series object
of class \code{ts}. It is useful when downstream functions require
a labeled series.
}
\examples{
ts_in <- ts(
  rnorm(12 * 30, mean = 10),
  start = c(1981, 1),
  frequency = 12
)

ts_labeled <- label_ts_mono(ts_in, label = "var")

}
