% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\encoding{UTF-8}
\name{lgssm}
\alias{lgssm}
\title{Function to estimate a linear Gaussian state-space model without exogenous variables}
\usage{
lgssm(ts_data)
}
\arguments{
\item{ts_data}{time series data}
}
\description{
Function to estimate a linear Gaussian state-space model without exogenous variables
}
