% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{logLik.tempssm}
\alias{logLik.tempssm}
\title{Log-likelihood method for tempssm objects (S3 method)}
\usage{
\method{logLik}{tempssm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tempssm"}, typically returned by
\code{tempssm()}.}

\item{...}{Additional arguments passed to the generic \code{logLik()} function.
These are currently ignored but are included for compatibility
with the generic interface.}
}
\value{
An object of class \code{"logLik"} containing the numeric log-likelihood,
with \code{df} and \code{nobs} attributes.
}
\description{
Log-likelihood method for tempssm objects (S3 method)
}
\examples{
\dontrun{
data(niigata_sst)

# fit model
res <- tempssm(niigata_sst)

# extract log-likelihood
ll <- logLik(res)

ll

# access attributes
attr(ll, "df") # number of parameters
attr(ll, "nobs") # number of observations
}

}
\seealso{
\code{\link{AIC.tempssm}} for computing AIC,
\code{\link{get_aic}} as a convenience wrapper.
}
