% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{monthly_anomaly}
\alias{monthly_anomaly}
\title{Calculate monthly temperature anomalies}
\usage{
monthly_anomaly(temp_ts)
}
\arguments{
\item{temp_ts}{Monthly temperature time series of class \code{ts}.
The time series must have a frequency of 12 (monthly data).}
}
\value{
A \code{ts} object of monthly temperature anomalies.
}
\description{
Monthly temperature anomalies are calculated by subtracting the
long-term monthly climatology from each observation.
}
\details{
Monthly climatology is computed using \code{mean_seasonal_cycle()}.
Missing values are ignored when calculating climatological means.
}
\examples{
temp_ts <- ts(
  c(5.2, 6.1, 9.3, 13.5, 17.8, 21.0,
    24.3, 25.1, 22.0, 17.1, 11.2, 7.0),
  start = c(2000, 1),
  frequency = 12
)

temp_anom <- monthly_anomaly(temp_ts)
tapply(as.numeric(temp_anom), cycle(temp_anom), mean)

}
