% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\encoding{UTF-8}
\name{monthly_csv2ts}
\alias{monthly_csv2ts}
\title{Function to convert from data.csv to monthly time sieries object}
\usage{
monthly_csv2ts(csv)
}
\arguments{
\item{csv}{input csv file}
}
\description{
Function to convert from data.csv to monthly time sieries object
}
