% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_tempssm.R
\name{plot.tempssm}
\alias{plot.tempssm}
\title{Plot method for tempssm objects}
\usage{
\method{plot}{tempssm}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{tempssm()}.}

\item{...}{Additional arguments passed to \code{autoplot.tempssm()}.}
}
\value{
Invisibly returns the plotted \code{ggplot} object.
}
\description{
Default base R plot method for objects fitted by \code{tempssm()}.
This method delegates to \code{autoplot.tempssm()} so that \code{plot(res)}
displays the same component plots as \code{autoplot(res)}.
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)
plot(res)
plot(res, component = "level", ci = FALSE)
}
}
