% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\encoding{UTF-8}
\name{plot_drift_ci}
\alias{plot_drift_ci}
\title{Function to plot drift component with confidence interval}
\usage{
plot_drift_ci(res, ci_range = 0.95)
}
\arguments{
\item{res}{output of model with using lgssm()}
}
\description{
Function to plot drift component with confidence interval
}
