% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\encoding{UTF-8}
\name{plot_level_trend_season_ar}
\alias{plot_level_trend_season_ar}
\title{Function to plot estimated components: level, drift, seasonal, and auto-regression}
\usage{
plot_level_trend_season_ar(res)
}
\arguments{
\item{res}{output of model with using lgssm()}
}
\description{
Function to plot estimated components: level, drift, seasonal, and auto-regression
}
