% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagonostics.R
\name{plot_residual_diagnostics}
\alias{plot_residual_diagnostics}
\title{Plot residual diagnostics for tempssm models}
\usage{
plot_residual_diagnostics(res, save = FALSE, prefix = "residuals")
}
\arguments{
\item{res}{An object of class "tempssm".}

\item{save}{Logical; if TRUE, plots are saved.}

\item{prefix}{Character prefix for file names.}
}
\description{
Plot residual diagnostics for tempssm models
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)

model_diagnose <- diagnose_residuals(res)
plot_residual_diagnostics(model_diagnose)
}
}
