% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\encoding{UTF-8}
\name{plot_typical_seasonal_cycle}
\alias{plot_typical_seasonal_cycle}
\title{Function to plot seasonal trend of monthly temperature}
\usage{
plot_typical_seasonal_cycle(ts)
}
\arguments{
\item{ts}{temperature time series object}
}
\description{
Function to plot seasonal trend of monthly temperature
}
