% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.tempssm}
\alias{print.summary.tempssm}
\title{Print method for summary of tempssm objects}
\usage{
\method{print}{summary.tempssm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.tempssm}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Prints a human-readable summary of a fitted
linear Gaussian state-space model estimated by
\code{tempssm()}.
}
\details{
This method is automatically called when a
\code{summary.tempssm} object is printed.
}
