% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_tempssm.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\title{Re-export the ggplot2 autoplot generic}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}
}}

