% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_cross_validation.R
\name{rolling_origin_tsCV}
\alias{rolling_origin_tsCV}
\title{Prediction based on fitted model for time series cross-validation}
\usage{
rolling_origin_tsCV(folds, fold_ids = 1)
}
\arguments{
\item{folds}{A \code{list} of folds object returned by \code{ts_cv_folds()}.}

\item{fold_ids}{A \code{vector} of sequential IDs of folds for ts cross-validation.}
}
\value{
A \code{list} object of ts cross-validation output
}
\description{
This function predict with using test data for in ts cross-validation.
}
