% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sst_jma2ts}
\alias{sst_jma2ts}
\title{Retrieve daily mean sea-surface temperature as a monthly ts object from JMA}
\usage{
sst_jma2ts(sea_area_id = NULL)
}
\arguments{
\item{sea_area_id}{Numeric sea area ID. The default is NULL
For example, 138 corresponding to the coastal sea off southern Ibaraki.
A list of sea area IDs and their corresponding regions is available at:
\url{https://www.data.jma.go.jp/kaiyou/data/db/kaikyo/series/engan/eg_areano.html}}
}
\value{
A \code{ts} object of monthly mean sea-surface temperature
with a single column named \code{Temp}.
}
\description{
This function downloads publicly available daily mean
sea-surface temperature (SST) data for Japanese coastal waters
provided by the Japan Meteorological Agency (JMA),
and returns the monthly average data as a \code{ts} object.
}
\details{
The function retrieves a text-format dataset from the JMA website,
parses daily observations, and constructs a \code{zoo} object
with calendar dates as the index.
Missing values in the original dataset are represented as \code{NA}.
}
\examples{
\dontrun{
sst_138_ts <- sst_jma2ts(sea_area_id = 138)
head(sst_138_ts)
}

}
