% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tempssm}
\alias{summary.tempssm}
\title{Summary method for tempssm objects}
\usage{
\method{summary}{tempssm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tempssm"} returned by \code{tempssm()}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list containing model diagnostics and summaries.
}
\description{
Provides a concise summary of a fitted linear Gaussian
state-space model estimated by \code{tempssm()}.
}
\examples{
\dontrun{
data(niigata_sst)

# fit model
res <- tempssm(niigata_sst)

# compute summary
s <- summary(res)

s

# access components programmatically
s$AIC
s$variances
}

}
