% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ThermoSSM}
\alias{summary.ThermoSSM}
\title{Summary method for ThermoSSM objects}
\usage{
\method{summary}{ThermoSSM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list containing model diagnostics and summaries.
}
\description{
Provides a concise summary of a fitted linear Gaussian
state-space model estimated by \code{lgssm()}.
}
