% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagonostics.R
\name{wrapper_checkresiduals}
\alias{wrapper_checkresiduals}
\title{Residual diagnostics for ThermoSSM models}
\usage{
wrapper_checkresiduals(
  res,
  plot_resid_save = FALSE,
  resid_file_name = "checkresiduals.png",
  JB_test = FALSE,
  plot_qq_save = FALSE,
  qq_file_name = "qqplot.png"
)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}

\item{plot_resid_save}{Logical; if TRUE, the residual diagnostic plot is saved.}

\item{resid_file_name}{Character; file name for saving the residual diagnostic plot (default: "checkresiduals.png").}

\item{JB_test}{Logical; if TRUE, Jarque–Bera test is optionally execute.}

\item{plot_qq_save}{Logical; if TRUE, the Q-Q plot of residuals is saved.}

\item{qq_file_name}{Character; file name for saving the Q-Q plot (default: "qqplot.png").}
}
\value{
A named \code{list} containing results of residual tests.
}
\description{
Wrapper function for residual diagnostics using functions from
the forecast, stats, tseries, and moments packages.
}
