% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\encoding{UTF-8}
\name{zoo_daily2ts_monthly}
\alias{zoo_daily2ts_monthly}
\title{Convert dairy zoo object to monthly ts object}
\usage{
zoo_daily2ts_monthly(zoo, var = "Temp", na.rm = TRUE)
}
\arguments{
\item{zoo}{zoo object}
}
\description{
Convert dairy zoo object to monthly ts object
}
