% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{AIC.tempssm}
\alias{AIC.tempssm}
\title{AIC method for tempssm objects (S3 method)}
\usage{
\method{AIC}{tempssm}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{"tempssm"}, typically returned by
\code{tempssm()}.}

\item{...}{Additional arguments passed to the generic \code{AIC()} function.
These are currently ignored but are included for compatibility
with the generic interface.}

\item{k}{Numeric penalty coefficient for the number of parameters.
This argument is included for compatibility with
\code{\link[stats]{AIC}} but is not used in the \pkg{tempssm} method,
where the standard AIC definition (\code{k = 2}) is applied.}
}
\value{
A numeric value giving the AIC of the fitted \code{tempssm} model.
}
\description{
Compute the Akaike Information Criterion (AIC) for a model fitted
by \code{tempssm()}. This method extends the generic
\code{\link[stats]{AIC}} function.
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)
aic <- AIC(res)
}

}
