% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_tempssm.R
\name{autoplot.tempssm}
\alias{autoplot.tempssm}
\title{Autoplot method for tempssm objects}
\usage{
\method{autoplot}{tempssm}(object, component = NULL, ci = TRUE, ci_level = 0.95, ...)
}
\arguments{
\item{object}{An object returned by \code{tempssm()}.}

\item{component}{Character string specifying which component to plot.
One of \code{"level"}, \code{"drift"}, \code{"season"}, or \code{"ar1"}.
If NULL (default), all components are plotted.}

\item{ci}{Logical; if TRUE (default), pointwise confidence intervals are shown.}

\item{ci_level}{Numeric confidence level between 0 and 1.
Defaults to \code{0.95}.}

\item{...}{Additional arguments passed to the corresponding
\code{autoplot_*()} function.}
}
\value{
A \code{ggplot} object if a single component is requested,
or a \pkg{patchwork} object combining multiple plots.
}
\description{
Create diagnostic and component plots for a model fitted by
\code{tempssm()}. By default, all available components are plotted.
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)

# plot all components with 95\% confidence interval
autoplot(res)

# plot all components without 95\% confidence interval
autoplot(res, ci = FALSE)

# plot each of components
autoplot(res, component = "level", ci = FALSE)
}

}
