% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{extract_AIC}
\alias{extract_AIC}
\title{Extract the Akaike Information Criterion (AIC)}
\usage{
extract_AIC(res)
}
\arguments{
\item{res}{An object of class \code{"tempssm"} returned by \code{ssm()}.}
}
\value{
A numeric value representing the AIC of the fitted model.
}
\description{
This function computes the Akaike Information Criterion (AIC)
for a fitted \code{tempssm} model.
}
\examples{
\dontrun{
data(niigata_sst)
res <- ssm(niigata_sst)
model_AIC <- extract_AIC(res)
}
}
