% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{get_aic}
\alias{get_aic}
\title{Extract the Akaike Information Criterion (AIC)}
\usage{
get_aic(res)
}
\arguments{
\item{res}{An object of class \code{"tempssm"} returned by \code{tempssm()}.}
}
\value{
A single numeric value representing the AIC of the fitted model.
}
\description{
Compute the Akaike Information Criterion (AIC) for a fitted
\code{tempssm} model using the model log-likelihood and the
number of estimated parameters.
}
\details{
The number of parameters is determined from the optimization results
stored in the fitted model. If exogenous variables are included,
their coefficients are added to the parameter count.
}
\examples{
\dontrun{
data(niigata_sst)
res <- tempssm(niigata_sst)
aic <- get_aic(res)
}
}
