% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgssm.R
\name{lgssm}
\alias{lgssm}
\title{Wrapper function for fitting a linear Gaussian state-space model}
\usage{
lgssm(
  temp_data,
  exo_data = NULL,
  ar_order = 1,
  use_seasonal = TRUE,
  inits = NULL,
  maxit = NULL,
  reltol = NULL
)
}
\arguments{
\item{temp_data}{A temperature time series of class \code{ts}.
The \code{ts} object must be univariant.
The series can have any arbitrary frequency of 2 or higher.
For example, a frequency of 12 represents a monthly \code{ts} object.}

\item{exo_data}{A data set of exogenous variable(s) of class \code{ts}.
The series may have any arbitrary frequency of 2 or higher,
but it must be the same as that of \code{temp_data}.
The default is \code{NULL} when fitting a model without exogenous variables.}

\item{ar_order}{Integer specifying the order of the autoregressive (AR) 
component in the error structure (e.g., 2 for AR(2), 3 for AR(3)). 
Default to 1.}

\item{use_seasonal}{Logical; Should seasonal component be considered. 
Default to TRUE.}

\item{inits}{Optional numeric vector of initial parameter values.
If \code{NULL}, default values are used.}

\item{maxit}{Optional numeric of maximum iteration.
If \code{NULL}, default value of 5000 is used.}

\item{reltol}{Optional numeric of reltol.
If \code{NULL}, default value of 1e-16 is used.}
}
\value{
An object of class \code{"tempssm"}, a named list containing:
\describe{
  \item{model}{Fitted \code{SSModel} object.}
  \item{fit}{Results from \code{fitSSM}.}
  \item{kfs}{Kalman filtering and smoothing results from \code{KFS}.}
  \item{data_temp}{Temperature time series used for estimation.}
  \item{data_exo}{Time series of exogenous variable(s) used for estimation.}
  \item{ar_order}{Order of the autoregressive component.}
  \item{use_seasonal}{Logical; whether to include a seasonal component.}
  \item{call}{Matched function call.}
}
}
\description{
Wrapper function for fitting a linear Gaussian state-space model
}
\examples{
\dontrun{
data(niigata_sst)
res <- lgssm(niigata_sst)
summary(res)
}
}
