% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ts.R
\name{lm_ts}
\alias{lm_ts}
\title{Fit a linear model to temperature time series}
\usage{
lm_ts(temp_data)
}
\arguments{
\item{temp_data}{A temperature time series of class \code{ts}.
The series can have any arbitrary frequency of 2 or higher.
For example, a frequency of 12 represents a monthly \code{ts} object.}
}
\value{
An object of class \code{list} containing:
\describe{
  \item{model}{Fitted \code{SSModel} object.}
  \item{fit}{Result of \code{fitSSM}.}
  \item{kfs}{Kalman filtering and smoothing results from \code{KFS}.}
  \item{data}{Temperature time series used for estimation.}
  \item{call}{Matched function call.}
}
}
\description{
This function estimates a linear model for monthly temperature time series.
Univariate \code{ts} object is supported.
If multivariate, a column named \code{"Temp"} is used.
}
