% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{logLik.tempssm}
\alias{logLik.tempssm}
\title{Log-likelihood method for tempssm objects (S3 method)}
\usage{
\method{logLik}{tempssm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tempssm"}, typically returned by
\code{tempssm()}.}

\item{...}{Additional arguments passed to the generic \code{logLik()} function.
These are currently ignored but are included for compatibility
with the generic interface.}
}
\description{
Log-likelihood method for tempssm objects (S3 method)
}
