% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\encoding{UTF-8}
\name{monthly_csv2ts}
\alias{monthly_csv2ts}
\title{Convert a monthly temperature CSV file to a monthly time series object}
\usage{
monthly_csv2ts(csv)
}
\arguments{
\item{csv}{Path to a CSV file containing monthly temperature data.}
}
\description{
Convert a monthly temperature CSV file to a monthly time series object
}
\details{
The input CSV file must have the following format:

\preformatted{
Year,Month,Temp
2001,1,10.4
2001,2,8.2
2001,3,NA
2001,4,13.6
...
}
}
