% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\encoding{UTF-8}
\name{plot_drift_ci}
\alias{plot_drift_ci}
\title{Function to plot drift component with confidence interval}
\usage{
plot_drift_ci(res, ci_level = 0.95)
}
\arguments{
\item{res}{An object of class \code{"ThermoSSM"} returned by \code{lgssm()}.}

\item{ci_level}{Confidence level for the interval estimation.
Must be a numeric value between 0 and 1.
The default is \code{0.95}, corresponding to a 95\% confidence interval.}
}
\description{
Function to plot drift component with confidence interval
}
