% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\encoding{UTF-8}
\name{plot_temp_dev}
\alias{plot_temp_dev}
\title{Function to plot time series of monthly temperature and temperature deviation}
\usage{
plot_temp_dev(ts)

plot_temp_dev(ts)
}
\arguments{
\item{ts}{temperature time series object}
}
\description{
Function to plot time series of monthly temperature and temperature deviation

Function to plot time series of monthly temperature and temperature deviation
}
