% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_cross_validation.R
\name{rolling_origin_tsCV}
\alias{rolling_origin_tsCV}
\title{Prediction based on fitted model for time series cross-validation}
\usage{
rolling_origin_tsCV(folds, fold_ids = NULL)
}
\arguments{
\item{folds}{A \code{list} of fold objects returned by \code{ts_cv_folds()}.}

\item{fold_ids}{An integer vector specifying which folds to run.
Default is all folds.}
}
\value{
A \code{list} object of ts cross-validation output.
}
\description{
This function performs prediction using test data in time series cross-validation.
}
