% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zoo_daily2ts_monthly}
\alias{zoo_daily2ts_monthly}
\title{Convert a daily zoo object to a monthly \code{ts} object}
\usage{
zoo_daily2ts_monthly(zoo_obj, var = "Temp", na.rm = TRUE, na_prop_max = 1)
}
\arguments{
\item{zoo_obj}{A \code{zoo} object with daily observations.
The index must be of class \code{Date} or \code{POSIXt}.}

\item{var}{A character string specifying the name of the variable
to be aggregated (default: \code{"Temp"}).}

\item{na.rm}{Logical; should missing values be removed before averaging?}

\item{na_prop_max}{Maximum allowed proportion of NA values within a month.
If the proportion of missing values exceeds this threshold, the monthly
mean is set to NA. Default is \code{1} (no additional filtering).}
}
\value{
A monthly \code{ts} object with frequency = 12.
}
\description{
Convert a daily zoo object to a monthly \code{ts} object
}
\examples{
sst_138_zoo <- sst_jma2zoo(sea_area_id = 138)
sst_138_ts_monthly <- zoo_daily2ts_monthly(sst_138_zoo, na_prop_max = 0.3)
head(sst_138_ts_monthly)

}
